pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karaoker"
version = "0.1.0"
description = "Japanese karaoke subtitle generator."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-only" }
authors = [{ name = "yuygfgg" }]
dependencies = [
  "pykakasi>=2.3.0",
]

[project.optional-dependencies]
textgrid = [
  "praat-textgrids>=1.4.0",
]
dev = [
  "pytest>=7.0",
  "ruff>=0.5",
]

[project.scripts]
karaoker = "karaoker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
